Detail Komplain
CRUD detail kolain dari form komplain

+++ Modules/Komplain/Http/Controllers/KomplainController.php
@@ -23,18 +23,19 @@ class KomplainController extends Controller
     public function index()
     {
         $data = new \stdClass();
-        $data->komplain = Komplain::with(['waroeng','komplain_details'])->get();
+        $data->komplain = Komplain::with(['waroeng','komplain_details'])->orderby('komplain_id','asc')->paginate(10);
+        $data->waroeng = Waroeng::All()->count();
+        $data->detail_komplain = KomplainDetail::with(['tb_kategori','komplain'])->get();
+        $data->kategori = TbKategori::All();
 
         $data->komplain2 = Komplain::groupBy('waroeng_id')
             ->select('waroeng_id', DB::raw('count(waroeng_id) as total'))
             ->with(['waroeng'])->get();
         $no=1;
 
-        $data->waroeng = Waroeng::All()->count();
-        
         return view('komplain::index', compact(['data','no']));
     }
-
+    
     /**
      * Show the form for creating a new resource.
      * @return Response
@@ -57,14 +58,25 @@ class KomplainController extends Controller
     public function store(Request $request)
     {
         $komplain = new Komplain();
-
-        // $komplain->id_kategori = $request->id_kategori;
+        
         $komplain->waroeng_id = $request->waroeng_id;
         $komplain->media_koplain = $request->media_komplain;
         $komplain->isi_komplain = $request->isi_komplain;
         $komplain->tanggal_komplain = $request->tanggal_komplain;
         $komplain->waktu_komplain = $request->waktu_komplain;
+        // dd($komplain);
         $komplain->save();
+
+            $tag = $request->id_kategori;
+            foreach($tag as $tags){
+                $detail_komplain = new KomplainDetail();
+                $detail_komplain->komplain_id = $komplain->komplain_id;
+                $detail_komplain->id_kategori =$tags;
+                // dd($detail_komplain);
+                $detail_komplain->save();
+            }
+        
+
 
         Session::flash('success',' Komplain added successfully');
         return Redirect::route('komplain.index');
@@ -88,8 +100,8 @@ class KomplainController extends Controller
         $data = new \stdClass();
         $data->komplain = Komplain::find($id);
         $data->waroeng = Waroeng::all();
-        // $data->kategori = TbKategori::all();
-        $data->detail_komplain = KomplainDetail::all();
+        $data->kategori = TbKategori::all();
+        $data->detail_komplain = KomplainDetail::where('komplain_id', $id)->get();
         return view('komplain::edit', compact('data'));
     }
 
@@ -102,14 +114,30 @@ class KomplainController extends Controller
     {
         $komplain = Komplain::find($id);
 
-        // $komplain->id_kategori = $request->id_kategori;
         $komplain->waroeng_id = $request->waroeng_id;
         $komplain->media_koplain = $request->media_komplain;
-		$komplain->isi_komplain = $request->isi_komplain;
-		$komplain->tanggal_komplain = $request->tanggal_komplain;
+        $komplain->isi_komplain = $request->isi_komplain;
+        $komplain->tanggal_komplain = $request->tanggal_komplain;
         $komplain->waktu_komplain = $request->waktu_komplain;
         $komplain->save();
-
+        KomplainDetail::where([
+            ['komplain_id',$id]
+        ])->delete();
+        if ($request->id_kategori){
+            foreach ($request->id_kategori as $k) {
+                $detail = KomplainDetail::where([
+                    ['komplain_id',$id],
+                ['id_kategori',$k]
+                ])->get();
+                
+                if ($detail){
+                    $d = new KomplainDetail;
+                    $d->id_kategori = $k;
+                    $d->komplain_id = $id;
+                    $d->save();
+                }
+            }
+        }
         return redirect()->route('komplain.index');
     }
 
@@ -119,8 +147,11 @@ class KomplainController extends Controller
      */
     public function destroy($id)
     {
-        $komplain = Komplain::find($id);
-        $komplain->delete();
+        $detail = KomplainDetail::where('komplain_id',$id);
+        if ($detail->delete()){
+            $komplain = Komplain::find($id);
+            $komplain->delete();
+        }
         return redirect()->route('komplain.index');
     }
